[OPD] address review: move OPD into loss_hub + add loss unit tests
Address Yueming's review comments on PR #993:
- Move apply_opd_kl_to_advantages out of loss.py into a dedicated
  loss_hub/opd.py, keeping loss.py a lightweight entrypoint. loss.py imports it
  for compute_advantages_and_returns.
- Add tests/fast/backends/training_utils/loss/test_opd.py covering the decoupled
  OPD loss path (weighted reverse-KL applied to advantages, opd_reverse_kl metric,
  None short-circuit, and the length/shape guard rails). A direct unit test
  rather than a loss-snapshot case, since snapshots live in the external
  miles-artifacts repo and would need regeneration there.
